backend/requirements.txt
# Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# AI & LangChain
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0

# Database
pymongo>=4.6.0
motor>=3.3.0

# HTTP & Web
requests>=2.31.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0

# Vector Database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.0
langchain-text-splitters>=0.0.1

# Video/Audio Processing
moviepy>=1.0.3
soundfile>=0.12.0
numpy>=1.24.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Auth
bcrypt>=4.1.0

# Browser/Scraping
browser-cookie3>=0.19.0

# SQL (optional)
sqlalchemy>=2.0.0